+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Chirptext changelog
 
+## 15 Mar 2022
+
+- Fix `ttl.read_json()` bug
+- Fix `token.surface` related bug (`token.surface()` is a method, not a property)
+
 ## 20 May 2021
 
 - v >= 0.1.2: Fix `newline` and `encoding` keywords missing in `chio.write_csv()` and `chio.write_tsv()`